cuad_app/requirements.txt
# core retrieval
chromadb==0.6.3
sentence-transformers==3.4.1
rank-bm25==0.2.2
pymupdf==1.25.3

# reranker
torch==2.6.0
transformers==4.50.0

# llm
google-generativeai==0.8.4

# orchestration
langchain==0.3.21
langchain-community==0.3.20
langchain-google-genai==2.0.10

# ui
gradio==5.23.1

# utils
python-dotenv==1.0.1
tqdm==4.67.1
numpy==1.26.4